refactor(transport): extract BaseHttpHost — close the sse↔http duplication (#158)
Issue #158 — `src/transport/sse.ts` and `src/transport/http.ts` were
~70% byte-identical: identical constructor, lifecycle (start/stop with
drain), CORS preflight, /health, origin allow-list, bearer-token auth,
shutdown gate, access logging, sessionCount + notify (added in #157
step 4), and per-session close. The two files only diverged at the
path-routing step.

This PR puts the shared scaffolding in a new abstract base:

  src/transport/base-http-host.ts (new) — BaseHttpHost<TTransport>:
    • Constructor + auth-token Buffer prep.
    • start() / stop() with drain + per-session close.
    • dispatch() pipeline up through the auth/shutdown gates.
    • handlePreflight + handleHealth (CORS bits parameterized via hooks).
    • verifyBearer, writeAccessLog.
    • notify(), sessionCount.
    • closeEntry() shared by both stop() loops.
    • Subclass hooks: logPrefix, bannerLabel, corsAllowedMethods,
      corsAllowedHeaders, setExtraCorsResponseHeaders,
      handleAuthenticatedRequest.

  src/transport/sse.ts (rewritten) — SseHost extends
  BaseHttpHost<SSEServerTransport>. Owns the GET /sse + POST /messages
  routing (long-lived GET /sse intentionally does NOT increment
  inFlight, which would block drain).

  src/transport/http.ts (rewritten) — StreamableHttpHost extends
  BaseHttpHost<StreamableHTTPServerTransport>. Owns POST/GET/DELETE
  /mcp routing, Mcp-Session-Id parsing, JSON body reading, and the
  JSON-RPC error envelope (HTTP-only).

Sizes:
  before: sse.ts 472 + http.ts 538 = 1010 lines
  after:  base-http-host.ts 441 + sse.ts 139 + http.ts 283 = 863
  net:    -147 lines (-15%) AND no more parallel copies that drift.

Behaviour preserved verbatim:
- Same gate ordering (preflight → health → origin → bearer → shutdown
  → CORS-response → routing).
- Same response shapes for every gate (403 'Origin not allowed', 401
  WWW-Authenticate, 503 + Retry-After, plain-text 404 for unknown
  paths in both transports, JSON-RPC envelope only for /mcp routes).
- Same long-lived semantics (SSE GET /sse not counted in inFlight).
- Same access-log format and content.
- Same close-on-stop semantics + the snapshot-then-iterate defence
  against `transport.onclose` deletes mid-iteration.

530/530 tests pass — the existing endpoint tests in `transport/sse.
test.ts` (live HTTP roundtrips) and `transport/http.test.ts` (real
MCP Client roundtrips) cover the full wire surface end-to-end.
sessionCount + notify regression tests (added in #157 step 4) still
hold; no new test file needed.

Closes #158.
